Add some tests for kth largest element.

+++ data_structures/source/heaps/kth_largest_element_in_stream.h
@@ -31,6 +31,6 @@ public:
 
 private:
 	std::priority_queue<T, std::vector<T>, std::greater<>> queue_;
-	int k_;
+	size_t k_;
 };
 #endif //KTH_LARGEST_ELEMENT_IN_STREAM_H
